+++ lib/cinch/plugin.rb
@@ -1,3 +1,5 @@
+require "cinch/helpers"
+
 # TODO more details in "message dropped" debug output
 module Cinch
   # This class represents the core of the plugin functionality of
@@ -79,7 +81,15 @@ module Cinch
       # @attr [Boolean] use_suffix
       # @attr [Symbol] method
       # @attr [Symbol] group
-      Matcher = Struct.new(:pattern, :use_prefix, :use_suffix, :method, :group, :prefix, :suffix, :react_on)
+      Matcher = Struct.new(:pattern,
+                           :use_prefix,
+                           :use_suffix,
+                           :method,
+                           :group,
+                           :prefix,
+                           :suffix,
+                           :react_on,
+                           :strip_colors)
 
       # Represents a Listener as created by {#listen_to}.
       #
@@ -105,7 +115,7 @@ module Cinch
       # @attr [Symbol] type
       # @attr [Array<Symbol>] for
       # @attr [Symbol] method
-      Hook = Struct.new(:type, :for, :method)
+      Hook = Struct.new(:type, :for, :method, :group)
 
       # @api private
       def self.extended(by)
@@ -179,14 +189,32 @@ module Cinch
       # @option options [Symbol, Fixnum] react_on (:message) The
       #   {file:docs/events.md event} to react on.
       # @option options [Symbol] :group (nil) The group the match belongs to.
+      # @option options [Boolean] :strip_colors (false) Strip colors
+      #   from message before attempting match
       # @return [Matcher]
       # @todo Document match/listener grouping
       def match(pattern, options = {})
-        options = {:use_prefix => true, :use_suffix => true, :method => :execute, :group => nil, :prefix => nil, :suffix => nil, :react_on => nil}.merge(options)
+        options = {
+          :use_prefix => true,
+          :use_suffix => true,
+          :method => :execute,
+          :group => nil,
+          :prefix => nil,
+          :suffix => nil,
+          :react_on => nil,
+          :strip_colors => false,
+        }.merge(options)
         if options[:react_on]
           options[:react_on] = options[:react_on].to_s.to_sym
         end
-        matcher = Matcher.new(pattern, *options.values_at(:use_prefix, :use_suffix, :method, :group, :prefix, :suffix, :react_on))
+        matcher = Matcher.new(pattern, *options.values_at(:use_prefix,
+                                                          :use_suffix,
+                                                          :method,
+                                                          :group,
+                                                          :prefix,
+                                                          :suffix,
+                                                          :react_on,
+                                                          :strip_colors))
         @matchers << matcher
 
         matcher
@@ -253,12 +281,15 @@ module Cinch
       #   a handler?
       # @option options [Array<:match, :listen_to, :ctcp>] :for ([:match, :listen_to, :ctcp])
       #   Which kinds of events to run the hook for.
-      # @option options [Symbol] :method (true) The method to execute.
+      # @option options [Symbol] :method (:hook) The method to execute.
+      # @option options [Symbol] :group (nil) The match group to
+      #   execute the hook for. Hooks belonging to the `nil` group
+      #   will execute for all matches.
       # @return [Hook]
       # @since 1.1.0
       def hook(type, options = {})
-        options = {:for => [:match, :listen_to, :ctcp], :method => :hook}.merge(options)
-        hook = Hook.new(type, options[:for], options[:method])
+        options = {:for => [:match, :listen_to, :ctcp], :method => :hook, :group => nil}.merge(options)
+        hook = Hook.new(type, options[:for], options[:method], options[:group])
         __hooks(type) << hook
 
         hook
@@ -266,7 +297,7 @@ module Cinch
 
       # @return [Hash]
       # @api private
-      def __hooks(type = nil, events = nil)
+      def __hooks(type = nil, events = nil, group = nil)
         if type.nil?
           hooks = @hooks
         else
@@ -280,14 +311,16 @@ module Cinch
           if hooks.is_a?(Hash)
             hooks = hooks.map { |k, v| v }
           end
-          return hooks.select { |hook| (events & hook.for).size > 0 }
+          hooks.select! { |hook| (events & hook.for).size > 0 }
         end
+
+        return hooks.select { |hook| hook.group.nil? || hook.group == group }
       end
 
       # @return [Boolean] True if processing should continue
       # @api private
-      def call_hooks(type, event, instance, args)
-        stop = __hooks(type, event).find { |hook|
+      def call_hooks(type, event, group, instance, args)
+        stop = __hooks(type, event, group).find { |hook|
           # stop after the first hook that returns false
           if hook.method.respond_to?(:call)
             instance.instance_exec(*args, &hook.method) == false
@@ -314,9 +347,9 @@ module Cinch
       self.class.listeners.each do |listener|
         @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Registering listener for type `#{listener.event}`"
         new_handler = Handler.new(@bot, listener.event, Pattern.new(nil, //, nil)) do |message, *args|
-          if self.class.call_hooks(:pre, :listen_to, self, [message])
+          if self.class.call_hooks(:pre, :listen_to, nil, self, [message])
             __send__(listener.method, message, *args)
-            self.class.call_hooks(:post, :listen_to, self, [message])
+            self.class.call_hooks(:post, :listen_to, nil, self, [message])
           else
             @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Dropping message due to hook"
           end
@@ -332,9 +365,9 @@ module Cinch
       self.class.ctcps.each do |ctcp|
         @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Registering CTCP `#{ctcp}`"
         new_handler = Handler.new(@bot, :ctcp, Pattern.generate(:ctcp, ctcp)) do |message, *args|
-          if self.class.call_hooks(:pre, :ctcp, self, [message])
+          if self.class.call_hooks(:pre, :ctcp, nil, self, [message])
             __send__("ctcp_#{ctcp.downcase}", message, *args)
-            self.class.call_hooks(:post, :ctcp, self, [message])
+            self.class.call_hooks(:post, :ctcp, nil, self, [message])
           else
             @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Dropping message due to hook"
           end
@@ -370,7 +403,11 @@ module Cinch
 
         @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Registering executor with pattern `#{pattern_to_register.inspect}`, reacting on `#{react_on}`"
 
-        new_handler = Handler.new(@bot, react_on, pattern_to_register, group: matcher.group) do |message, *args|
+        new_handler = Handler.new(@bot,
+                                  react_on,
+                                  pattern_to_register,
+                                  group: matcher.group,
+                                  strip_colors: matcher.strip_colors) do |message, *args|
           method = method(matcher.method)
           arity = method.arity - 1
           if arity > 0
@@ -378,9 +415,9 @@ module Cinch
           elsif arity == 0
             args = []
           end
-          if self.class.call_hooks(:pre, :match, self, [message])
+          if self.class.call_hooks(:pre, :match, matcher.group, self, [message])
             method.call(message, *args)
-            self.class.call_hooks(:post, :match, self, [message])
+            self.class.call_hooks(:post, :match, matcher.group, self, [message])
           else
             @bot.loggers.debug "[plugin] #{self.class.plugin_name}: Dropping message due to hook"
           end
@@ -454,8 +491,8 @@ module Cinch
     end
 
     # (see Bot#synchronize)
-    def synchronize(*args, &block)
-      @bot.synchronize(*args, &block)
+    def synchronize(name, &block)
+      @bot.synchronize(name, &block)
     end
 
     # Provides access to plugin-specific options.